Create README.md
Welcome to the CloudSimEx project!

The goal of the CloudSimEx project is to develop a set of extensions for the CloudSim simulator. The extensions that prove to be worthy will be merged with CloudSim later on.

Note! These extensions are not officially supported by the CloudSim team until they are integrated with CloudSim.

Currently CloudSimEx features:
  * Web session modelling;
  * Better logging utilities;
  * Utilities for generating CSV files for statistical analysis;
  * Automatic id generation;
  * Utilities for running multiple experiments in parallel;
  * Utilities for modelling network latencies;
  * MapReduce simulation.

Please read through the wiki at - https://github.com/Cloudslab/CloudSimEx/wiki